src/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
graphene-django==3.1.5
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
django-filter==23.5
psycopg[binary]==3.2.1
PyJWT==2.9.0
djangorestframework-simplejwt==5.3.1

